requirements.txt
filelock==3.10.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
six==1.16.0
sympy==1.11.1
torch==2.0.0
torch-vision==0.1.6.dev0
typing_extensions==4.5.0
